feat: Add status field to CreateModel and UpdateModel functions
The CreateModel and UpdateModel functions in the cruds/model.py file have been updated to include a new status field. This field allows for enabling or disabling a model. When a model is enabled, all other models with the same status will be disabled. This provides better control and management of the models.